% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approxLeadTimes.R
\name{approxLeadTimes}
\alias{approxLeadTimes}
\title{RAnEnExtra::approxLeadTimes}
\usage{
approxLeadTimes(
  forecasts,
  flt_dimension,
  old_flts,
  new_flts,
  method = "linear",
  parallel = FALSE
)
}
\arguments{
\item{forecasts}{A 4-dimensional array}

\item{flt_dimension}{Which dimension is the lead time dimension.}

\item{old_flts}{Original forecast lead times in seconds}

\item{new_flts}{New forecast lead times in seconds}

\item{method}{The \code{method} argument passed to \code{approx}.}

\item{parallel}{Whether to use \code{future_apply} for parallelization}
}
\description{
RAnEnExtra::approxLeadTimes interpolate forecasts across the lead time dimension.
}
\author{
Weiming Hu \email{weiming@psu.edu}

Martina Calovi \email{mxc895@psu.edu}
}
