% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasCorrection.R
\name{biasCorrection}
\alias{biasCorrection}
\title{RAnEnExtra::biasCorrection}
\usage{
biasCorrection(
  analogs,
  target.forecasts,
  historical.forecasts,
  similarity.time.index,
  similarity.station.index = NULL,
  regression.forecasts = NULL,
  regression.observations = NULL,
  forecast.id = NULL,
  activation.func = NULL,
  show.progress = T,
  return.more = F,
  group.func = mean,
  ...
)
}
\arguments{
\item{analogs}{A four-dimensional array for analog values with the dimensions \verb{[stations, test times, lead times, members]}}

\item{target.forecasts}{A three- or four- dimensional array for test forecasts that analogs are generated for. The dimensions
can either be \verb{[stations, test times, lead times]} or \verb{[stations, test times, lead times, 1]}.}

\item{historical.forecasts}{The historical forecast search repository with the dimensions \verb{[parameters, stations, times, lead times]}.}

\item{similarity.time.index}{Similarity time index for each analog members}

\item{similarity.station.index}{Similarity station index for each analog members}

\item{regression.forecasts}{The forecast values used to calculate the slope of a linear regression line. These forecasts must
correspond to observations for regression.}

\item{regression.observations}{The observation values used to calculate the slope of a linear regression line These
observations must correspond to forecasts for regression.}

\item{forecast.id}{A forecasts parameter index used by the \code{historical.forecasts} to specify which forecast parameter to use.}

\item{activation.func}{An activation function to signify whether a particular ensemble should be bias corrected. This function
should takes a single argument and return a TRUE or FALSE. The single argument of the function will be an analog ensemble,
or a numeric vector. For example, \code{activation.func = function(members) {if (mean(members) > 15) {return(T)} else {return(F)}}}
will only correct ensembles that have an average over 15.}

\item{show.progress}{Whether to show a progress bar}

\item{return.more}{Whether to return more information}

\item{group.func}{How to collapse the analog ensemble to a single value to calculate the amount of correction. In the paper and
by default, this is \code{mean}.}

\item{...}{Additional variables passed to \code{group.func}}
}
\description{
RAnEnExtra::biasCorrection carries out a bias correction routine on the analog ensembles
using a linear regression method. This correction method is useful for predicting extreme event.
}
\references{
Alessandrini, Stefano, Simone Sperati, and Luca Delle Monache. "Improving the analog ensemble
wind speed forecasts for rare events." Monthly Weather Review 147.7 (2019): 2677-2692.
}
\author{
Weiming Hu \email{weiming@psu.edu}

Martina Calovi \email{mxc895@psu.edu}
}
