% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut.pixels.along.y.R
\name{cut.pixels.along.y}
\alias{cut.pixels.along.y}
\title{divide pixels according to cuts on y}
\usage{
cut.pixels.along.y(pixels, ycuts, xgrids.total, ygrids.total, start,
  flag.sort = T)
}
\arguments{
\item{pixels}{a vector of indices of pixels that
will get divided}

\item{ycuts}{a vector of integers specifing the y
locations that will be cut. The range of the cut will
be defined as [ycuts[i], ycuts[i+1]).}

\item{xgrids.total}{total number of x.}

\item{ygrids.total}{total number of y.}

\item{start}{the counting start of pixels.}

\item{flag.sort}{a bool specifing whether to sort
the cut pixels or not}
}
\value{
a list of vectors with the separate pixels.
}
\description{
cut.pixels.along.y cuts the pixels into separate
sets based on the cutting points specified in
ycuts and the y coordinates of each pixel.
The start counting index of pixels should
be specified in the argument start. Note that
ycuts should be counted from 1
}
