% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.points.in.triangle.R
\name{get.points.in.triangle}
\alias{get.points.in.triangle}
\title{get the x and y of points within the triangle}
\usage{
get.points.in.triangle(triangle, bound = T)
}
\arguments{
\item{triangle}{a spatialpolygon object of a triangle}

\item{bound}{If points fall exactly on polygon boundaries,
the default NULL gives arbitrary assignments. If TRUE,
then all points "on" boundaries are set as within the
polygon, if FALSE, outside.}
}
\value{
a matrix with x and y of points in the triangle
}
\description{
get the x and y of points within the triangle
}
