% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include.boundaries.R
\name{include.boundaries}
\alias{include.boundaries}
\title{include.boundaries select the points on the four
edges randomly and replace the original points
in pixels.}
\usage{
include.boundaries(pixels, num.edge.points, xgrids.total, ygrids.total, start)
}
\arguments{
\item{pixels}{a vector of indices of pixels that
will get divided}

\item{num.edge.points}{an integer specifies the
extra points to select on each edge}

\item{xgrids.total}{total number of x.}

\item{ygrids.total}{total number of y.}

\item{start}{the counting start of pixels.}
}
\value{
same object as the argument pixels but
with the pixels on the boundaries
}
\description{
include.boundaries select the points on the four
edges randomly and replace the original points
in pixels.
}
