% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include.corners.R
\name{include.corners}
\alias{include.corners}
\title{check and include the pixels at the 4 corners}
\usage{
include.corners(pixels, xgrids.total, ygrids.total, start)
}
\arguments{
\item{pixels}{a vector of indices of pixels that
will get divided}

\item{xgrids.total}{total number of x.}

\item{ygrids.total}{total number of y.}

\item{start}{the counting start of pixels.}
}
\value{
same object as the argument pixels but
being sure the 4 corner pixels are included
}
\description{
include.corners will return the same vector if the corner
pixels are already included. Otherwise, it will replace
    first element   -->   left top pixel
    second element  -->   right top pxel
    third element   -->   left bottom pixel
    forth element   -->   right bottom pixel
}
\details{
Pixels are counted row-wise. There should at least be
4 pixels; otherwise nothing will be done. The start
counting index of pixels should be specified in the
argument start.
}
