% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse.distance.interpolation.R
\name{inverse.distance.interpolation}
\alias{inverse.distance.interpolation}
\title{interpolate values using inverse Euclidean distance weighting}
\usage{
inverse.distance.interpolation(pts.interpolate, pts.known, pts.values)
}
\arguments{
\item{pts.interpolate}{coordinates for points that will be
interpolated}

\item{pts.known}{coordinates for known points}

\item{pts.values}{values for known points}
}
\value{
a vector with values for the points to be
interpolated
}
\description{
interpolate values using inverse Euclidean distance weighting
}
