% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeAnalogsByStations.R
\name{mergeAnalogsByStations}
\alias{mergeAnalogsByStations}
\title{RAnEnExtra::mergeAnalogsByStations}
\usage{
mergeAnalogsByStations(
  files,
  verbose = T,
  merge_forecasts = T,
  cores = 1,
  copy_vars = NULL
)
}
\arguments{
\item{files}{A vector of AnEn files}

\item{verbose}{Whether to be verbose}

\item{merge_forecasts}{Whether to merge \code{Forecasts/Data} if it exists}

\item{cores}{The number of cores to use while reading files}

\item{copy_vars}{The variable names to simply copy from the first file in the (sorted) list into the final AnEn list}
}
\value{
An AnEn.
}
\description{
RAnEnExtra::mergeAnalogsByStations reads a vector of analog files, usually generated from \code{anen_netcdf} or \code{anen_grib},
and then combines analogs by stations. It is assumed that \code{analogs}, \code{Xs}, and \code{Ys} exist in all files.
}
