% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nni.R
\name{nni}
\alias{nni}
\title{nearest neighbor interpolation}
\usage{
nni(x, y, z, rast, n = 3)
}
\arguments{
\item{x}{a vector of x coordinates}

\item{y}{a vector of y coordinates}

\item{z}{a vector of values of each point}

\item{rast}{a base raster that specifies the outline
of the raster to be interpolated}

\item{n}{number of nearest neighbors}
}
\value{
an interpolated raster object
}
\description{
nni interpolate a raster based on the given control
points using nearest neighbor interpolation.
}
