% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizeBinnedSpreadSkill.R
\name{organizeBinnedSpreadSkill}
\alias{organizeBinnedSpreadSkill}
\title{RAnEnExtra::organizeBinnedSpreadSkill}
\usage{
organizeBinnedSpreadSkill(results)
}
\arguments{
\item{results}{The results from RAnEnExtra::verify. It should be a list. This function can only
parse verification generated by the metric \code{BinnedSpreadSkill}; any other metrics will be ignored.}
}
\value{
A data frame
}
\description{
RAnEnExtra::organizeBinnedSpreadSkill organize the verification results into a flat table format
so that it is easy to plot with ggplot.
}
\author{
Weiming Hu \email{weiming@psu.edu}
}
