% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizeFLTs.R
\name{organizeFLTs}
\alias{organizeFLTs}
\title{RAnEnExtra::organizeFLTs}
\usage{
organizeFLTs(results, flts, boot = FALSE, parse_metrics = NULL)
}
\arguments{
\item{results}{The results from RAnEnExtra::verify. It should be a list.}

\item{flts}{The forecast lead times.}

\item{boot}{Whether \code{boot} is used during the verification process}

\item{parse_metrics}{The metrics to organize into a data frame. Default to all supported metrics.}
}
\value{
a data frame or a data table
}
\description{
RAnEnExtra::organizeFLTs organize the verification results into a flat table format so that
it is easy to plot with ggplot. Currently, this function only knows how to organize a portion
of the verification metrics.
}
\author{
Weiming Hu \email{weiming@psu.edu}
}
