% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseProfiling.R
\name{parseProfiling}
\alias{parseProfiling}
\title{RAnEnExtra::parseProfiling}
\usage{
parseProfiling(file, first_up = T)
}
\arguments{
\item{file}{The file to parse}

\item{first_up}{Convert the first letter in the name column to uppercase.}
}
\value{
A data frame with parsed messages as columns
}
\description{
RAnEnExtra::parseProfiling is used to parse standard output messages generated
from the commandline utilities, e.g. \code{anen_grib} and \code{anen_grib_mpi}
(\href{https://github.com/Weiming-Hu/AnalogsEnsemble/tree/master/apps}{What are these}?
\href{https://weiming-hu.github.io/AnalogsEnsemble/doc#canen}{How to install these}?).
}
\author{
Weiming Hu \email{weiming@psu.edu}
}
