% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixels.to.x.by.row.R
\name{pixels.to.x.by.row}
\alias{pixels.to.x.by.row}
\title{convert pixel index to x coordinate}
\usage{
pixels.to.x.by.row(pixels, xgrids.total, start)
}
\arguments{
\item{pixels}{a vector of indices of pixels that
will get divided}

\item{xgrids.total}{total number of x.}

\item{start}{the counting start of pixels.}
}
\value{
x coordinate(s)
}
\description{
convert the pixel index to its x coordinate counting
row-wise. The start matters because it sets the offset.
The returned numbers are counted from 1.
}
