% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixelsFromXs.R
\name{pixelsFromXs}
\alias{pixelsFromXs}
\title{RAnEnExtra::pixelsFromXs}
\usage{
pixelsFromXs(x, y, xgrids.total, start)
}
\arguments{
\item{x}{a vector of x coordinates of the pixels
starting from 1}

\item{y}{a vector of y coordinates of the pixels
starting from 1}

\item{xgrids.total}{total number of x.}

\item{start}{the counting start of pixels.}
}
\value{
the index of the pixel
}
\description{
RAnEnExtra::pixelsFromXs convert x and y to pixel index
}
\details{
convert x and y coordinates index to pixels index. Assume
that x and y coordinates index are counting from 1. Pixels
start counting from the argument start, and they are
counted row-wise.
}
