% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixelsSplitByY.R
\name{pixelsSplitByY}
\alias{pixelsSplitByY}
\title{RAnEnExtra::pixelsSplitByY}
\usage{
pixelsSplitByY(
  pixels,
  ycuts,
  xgrids.total,
  ygrids.total,
  start = 1,
  flag.sort = T
)
}
\arguments{
\item{pixels}{a vector of indices of pixels that
will get divided}

\item{ycuts}{a vector of integers specifing the y
locations that will be cut. The range of the cut will
be defined as [ycuts[i], ycuts[i+1]).}

\item{xgrids.total}{total number of x.}

\item{ygrids.total}{total number of y.}

\item{start}{the counting start of pixels.}

\item{flag.sort}{a bool specifing whether to sort
the cut pixels or not}
}
\value{
a list of vectors with the separate pixels.
}
\description{
RAnEnExtra::pixelsSplitByY divide pixels according
to cuts on along the y axis.
}
\details{
cut.pixels.along.y cuts the pixels into separate
sets based on the cutting points specified in
ycuts and the y coordinates of each pixel.
The start counting index of pixels should
be specified in the argument start. Note that
ycuts should be counted from 1
}
