% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixelsToX.R
\name{pixelsToX}
\alias{pixelsToX}
\title{RAnEnExtra::pixelsToX}
\usage{
pixelsToX(pixels, xgrids.total, start)
}
\arguments{
\item{pixels}{a vector of indices of pixels that
will get divided}

\item{xgrids.total}{total number of x.}

\item{start}{the counting start of pixels.}
}
\value{
x coordinate(s)
}
\description{
RAnEnExtra::pixelsToX convert the pixel index
to its x coordinate counting row-wise. The start matters
because it sets the offset. The returned numbers are
counted from 1.
}
