% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixelsToY.R
\name{pixelsToY}
\alias{pixelsToY}
\title{RAnEnExtra::pixelsToY}
\usage{
pixelsToY(pixels, xgrids.total, start)
}
\arguments{
\item{pixels}{a vector of indices of pixels that
will get divided}

\item{xgrids.total}{total number of x.}

\item{start}{the counting start of pixels.}
}
\value{
y coordinate(s)
}
\description{
RAnEnExtra::pixelsToY convert the pixel index
to its y coordinate counting row-wise. The start matters
because it sets the offset. The returned numbers are
counted from 1.
}
