% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAnalogSample.R
\name{plotAnalogSample}
\alias{plotAnalogSample}
\title{RAnEnExtra::plotAnalogSample}
\usage{
plotAnalogSample(forecasts, fcst.id, fcst.times, flts, observations,
  obs.id, obs.times, analogs, sims, sims.index, test.times, i.station,
  i.test.day, num.analogs, weights, parameter.names,
  share.parameter.names = F, flts.fraction = 3600,
  boxplot.outliers = F, border.anen.box = "grey",
  lty.current.fcsts = "longdash", col.current.fcsts = "red",
  cex.current.fcst = 1, pch.current.obs = 1,
  col.current.obs = "black", cex.current.obs = 1.5, cex.label = 1.5,
  legend.texts = c("Observations", "NAM"), legend.position = "topleft",
  single.figure = T, mar = c(1.5, 5, 1, 1) + 0.1, omi = c(0.3, 0, 0,
  0))
}
\arguments{
\item{forecasts}{The forecast data array}

\item{fcst.id}{The index in forecast parameters that which parameter is predicted.}

\item{fcst.times}{Forecast times}

\item{flts}{Forecast lead times}

\item{observations}{Observation data array}

\item{obs.id}{Observation variable id}

\item{obs.times}{Observation times}

\item{analogs}{The generated analogs with 4 dimensions.}

\item{sims}{Similarity values}

\item{sims.index}{Similarity time index}

\item{test.times}{Test times for analogs}

\item{i.station}{The station index from analogs.}

\item{i.test.day}{The test day index from analogs.}

\item{num.analogs}{The number of analogs.}

\item{weights}{The weights used in analog generation.}

\item{parameter.names}{The forecast parameter names.}

\item{share.parameter.names}{Whether forecasts and observations have the same
parameters. If this is TRUE, the observations will be plot on the other parameters
that are used to compute the simialrity.}

\item{flts.fraction}{This is the fraction when dividing FLT values. 3600 means that
the x labels are in hours when FLTs are in seconds.}

\item{boxplot.outliers}{Whether to plot outliers.}

\item{border.anen.box}{The color of the box plots.}

\item{lty.current.fcsts}{The line type for current forecasts.}

\item{col.current.fcsts}{The color for current forecasts.}

\item{cex.current.fcst}{The size for current forecasts.}

\item{pch.current.obs}{The point type for current observations.}

\item{col.current.obs}{The color for current observations.}

\item{cex.current.obs}{The size for current observations.}

\item{cex.label}{The font size for x, y, and the legend labels.}

\item{legend.texts}{A vector of characters of length 2. The first one
is for observations, and the second one for forecasts.}

\item{legend.position}{The legend position.}

\item{single.figure}{Whether to generate a single plot with subplots.}

\item{mar}{Margin for the plot. This is effective only when single.figure is TURE.}

\item{omi}{Outer margin for the plot. This is effective only when single.figure is TURE.}

\item{member.name}{Which member from AnEn should be plotted. This is useful
when you have bias corrected analog members in AnEn that you would like to plot.}
}
\description{
RAnEnExtra::plotAnalogSample compares the current forecasts and AnEn with observations
for one particular station and test time across all FLTs available in forecasts.
It also plots the historical forecasts that are identified by AnEn, and
compares them with current forecasts for each parameters used to generate
analogs.
}
\details{
\strong{This function only works for AnEnIS.}
}
\author{
Weiming Hu \email{cervone@psu.edu}

Laura CLemente-Harding \email{lec170@psu.edu}

Martina Calovi \email{mcalovi@psu.edu}

Guido Cervone \email{cervone@psu.edu}
}
