% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryObservationsSpace.R
\name{queryObservationsSpace}
\alias{queryObservationsSpace}
\title{RAnEnExtra::queryObservationsSpace}
\usage{
queryObservationsSpace(
  analogs.index,
  analogs.x,
  analogs.y,
  analogs.flt,
  obs,
  obs.x,
  obs.y,
  obs.id,
  verbose = 3,
  show.progress = F,
  keep.station.table = F
)
}
\arguments{
\item{analogs.index}{A 4-dimensional array for analogs time index generated from
\href{https://weiming-hu.github.io/AnalogsEnsemble/R/reference/generateAnalogs.html}{RAnEn::generateAnalogs}.}

\item{analogs.x}{The x coordinates for analog stations.}

\item{analogs.y}{The y coordinates for analog stations.}

\item{analogs.flt}{The forecast lead times for analog stations in number of seconds.}

\item{obs}{A 3-dimensional array for observations used to generate
analogs. It is usually the member \code{search_observations} from \code{config} that
is used by \href{https://weiming-hu.github.io/AnalogsEnsemble/R/reference/generateAnalogs.html}{RAnEn::generateAnalogs}.}

\item{obs.x}{The x coordinates for observation stations.
These will be the target stations that analogs will be downscaled to.}

\item{obs.y}{The y coordinates for observation stations.
These will be the target stations that analogs will be downscaled to.}

\item{obs.id}{The observation variable ID that will be used to downscale.}

\item{verbose}{Verbose level for output messages.}

\item{show.progress}{Whether to show progress information.}

\item{keep.station.table}{Whether to keep the lookup table for matching
AnEn stations and the target observation stations in the return.}
}
\value{
An \code{AnEn} list
}
\description{
RAnEnExtra::queryObservationsSpace carries out the spatial downscaling
by reusing the indices from AnEn at different locations.
}
\author{
Weiming Hu \email{weiming@psu.edu}

Martina Calovi \email{mcalovi@psu.edu}

Guido Cervone \email{cervone@psu.edu}
}
