% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryObservationsTime.R
\name{queryObservationsTime}
\alias{queryObservationsTime}
\title{RAnEnExtra::queryObservationsTime}
\usage{
queryObservationsTime(
  analogs.index,
  analogs.flt,
  obs,
  obs.id,
  obs.time,
  radius = 1.5 * 3600,
  res = 15 * 60,
  verbose = 3,
  show.progress = F,
  keep.downscaled.FLTs = F
)
}
\arguments{
\item{analogs.flt}{The forecast lead times for analog stations in number of seconds.}

\item{obs}{A 3-dimensional array for observations used to generate
analogs. It is usually the member \code{search_observations} from \code{config} that
is used by \href{https://weiming-hu.github.io/AnalogsEnsemble/R/reference/generateAnalogs.html}{RAnEn::generateAnalogs}.}

\item{obs.id}{The observation variable ID that will be used to downscale.}

\item{obs.time}{The obseravtion times. This is usually the member \code{observation_times}
from \code{config} that is used by
\href{https://weiming-hu.github.io/AnalogsEnsemble/R/reference/generateAnalogs.html}{RAnEn::generateAnalogs}.}

\item{radius}{The number of seconds that each forecast lead time will be extended to.}

\item{res}{The number of seconds for the downscaled tempoeral resolution.}

\item{verbose}{Verbose level for output messages.}

\item{show.progress}{Whether to show progress information.}

\item{keep.downscaled.FLTs}{Whether to return the downscaled forecast lead times.}

\item{analogs}{A 4-dimensional array for analogs time index generated from
\href{https://weiming-hu.github.io/AnalogsEnsemble/R/reference/generateAnalogs.default.html}{RAnEn::generateAnalogs}.}
}
\value{
An \code{AnEn} list.
}
\description{
RAnEnExtra::queryObservationsTime carries out the temporal downscaling by
reusing the indices from AnEn at forecast lead times. The downscaled time
series are defined with the original forecast lead time being at the center.
}
\author{
Weiming Hu \email{weiming@psu.edu}

Martina Calovi \email{mcalovi@psu.edu}

Guido Cervone \email{cervone@psu.edu}
}
