% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.points.in.triangle.R
\name{random.points.in.triangle}
\alias{random.points.in.triangle}
\title{randomly select points within the triangle}
\usage{
random.points.in.triangle(triangle, num.points, bound = T, verbose = F)
}
\arguments{
\item{triangle}{a spatialpolygon object of a triangle}

\item{num.points}{an integer specifying the amount of
random points to select}

\item{bound}{If points fall exactly on polygon boundaries,
the default NULL gives arbitrary assignments. If TRUE,
then all points "on" boundaries are set as within the
polygon, if FALSE, outside.}

\item{verbose}{a bool controls the standard output level}
}
\value{
a matrix with x and y coordinates of the random
points
}
\description{
random.points.in.trianglegon randomly select num.points
points in the triangular polygon, and return the points
}
