% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.vector.duplicates.R
\name{remove.vector.duplicates}
\alias{remove.vector.duplicates}
\title{removes the duplicated elements in a vector}
\usage{
remove.vector.duplicates(v)
}
\arguments{
\item{v}{a vector to check and remove the
duplicates}
}
\value{
a vector with no duplicates
}
\description{
remove.vector.duplicates removes the
duplicated elements from the original vector
and return the new vector with no duplicates
}
