% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicateDimension.R
\name{replicateDimension}
\alias{replicateDimension}
\title{RAnEnExtra::replicateDimension}
\usage{
replicateDimension(arr, times, dimension)
}
\description{
RAnEnExtra::replicateDimension replicates a particular dimensions of the arrray.
}
\author{
Weiming Hu \email{weiming@psu.edu}

Martina Calovi \email{mxc895@psu.edu}
}
