% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showVerificationMetrics.R
\name{showVerificationMetrics}
\alias{showVerificationMetrics}
\title{RAnEnExtra::showVerificationMetrics}
\usage{
showVerificationMetrics()
}
\description{
RAnEnExtra::showVerificationMetrics prints all the supported verification metrics.
}
\author{
Weiming Hu \email{weiming@psu.edu}
}
