% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetParameters.R
\name{subsetParameters}
\alias{subsetParameters}
\title{RAnEnExtra::subsetParameters}
\usage{
subsetParameters(index, l, verbose = T)
}
\arguments{
\item{index}{An index vector for parameters to extract}

\item{l}{A forecast or observation list. For how to create such a
list, please see
\href{https://weiming-hu.github.io/AnalogsEnsemble/2019/11/18/format-obs.html}{this tutorial}.
For what members to include in the list, see
\href{https://weiming-hu.github.io/AnalogsEnsemble/2019/01/16/NetCDF-File-Types.html}{this doc}.}

\item{verbose}{Whether to print progress information.}
}
\value{
A forecast or observation list depending on your input
list type with the subset parameters.
}
\description{
RAnEnExtra::subsetParameters is a convevient function to subset parameters
from forecast and observation lists.
}
\details{
RAnEnExtra::subsetParameters will select the parameters based on the input
index from the following members of the input list (if they exist):
\itemize{
\item ParameterNames
\item ParameterCirculars
\item Data
}
}
\seealso{
\link[RAnEnExtra:subsetStations]{RAnEnExtra::subsetStations} which shares a similar interface and has an usage example.
}
