% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament.selection.R
\name{tournament.selection}
\alias{tournament.selection}
\title{tournament.selection implements the tournament selection
algorithm}
\usage{
tournament.selection(fitness, num.samples, num.champions = 1,
  tournament.size = 2, replacement = F)
}
\arguments{
\item{fitness}{fitness value}

\item{num.samples}{number of samples to select}

\item{num.champions}{number of champions to choose from
each tournament}

\item{tournament.size}{number of candidates to choose
from the population}

\item{replacement}{whether to select samples with
replacement or not}
}
\value{
a vector with indices for the selected samples
}
\description{
tournament.selection implements the tournament selection
algorithm
}
