% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightSearch.R
\name{weightSearch}
\alias{weightSearch}
\title{RAnEnExtra::weightSearch}
\usage{
weightSearch(
  weights,
  forecasts,
  observations,
  test.times,
  search.times,
  config,
  metric = "RMSE",
  return.best.only = F
)
}
\arguments{
\item{weights}{Either a numeric matrix or a single numeral. If it is a matrix, each row is a certain weight combination and the number
of rows equal to the number of iterations; if it is a single numeral, it is the number of iterations and the weights will be randomly generated.}

\item{forecasts}{Forecasts for \code{RAnEn::generateAnalogs.Forecasts}}

\item{observations}{Observations for \code{RAnEn::generateAnalogs.Forecasts}}

\item{test.times}{Test times for \code{RAnEn::generateAnalogs.Forecasts}}

\item{search.times}{Search times for \code{RAnEn::generateAnalogs.Forecasts}}

\item{config}{Config for \code{RAnEn::generateAnalogs.Forecasts}}

\item{metric}{The metric for \code{RAnEnExtra::verify}. Please note two things: (1) please make sure the member \code{mean} exists in the verification
results because it is used to evaluate different combinations of weights; (2) it is assumed that the metric is an error measurement so that
the lowest metric is considered the best.}

\item{return.best.only}{Whether to only return the best combination of weights}
}
\value{
Either a vector or a matrix. If a vector is returned, it is the best combination of weights; if a matrix is returned, it is the
combinations of weights that have been searched and the last column is the verification metric.
}
\description{
RAnEnExtra::weightSearch is a grid search function for weight optimization.
}
\author{
Alon Sidel \email{mxc895@psu.edu}

Weiming Hu \email{cervone@psu.edu}
}
