% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeAnalogsByStations.R
\name{mergeAnalogsByStations}
\alias{mergeAnalogsByStations}
\title{RAnEnExtra::mergeAnalogsByStations}
\usage{
mergeAnalogsByStations(files, verbose = T)
}
\arguments{
\item{files}{A vector of AnEn files}

\item{verbose}{Whether to be verbose}
}
\value{
An AnEn.
}
\description{
RAnEnExtra::mergeAnalogsByStations reads a vector of analog files, usually generated from \code{anen_netcdf} or \code{anen_grib},
and then combines analogs by stations. It is assumed that \code{analogs}, \code{Xs}, and \code{Ys} exist in all files.
}
